pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-project"
version = "1.0.0"
description = "An intelligent AI agent system built with LangGraph, FastAPI, and Azure OpenAI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "agent", "langgraph", "fastapi", "streamlit", "azure-openai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.1",
    "langchain>=0.1.4",
    "langchain-community>=0.0.13",
    "langchain-core>=0.1.12",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "openai>=1.12.0",
    "azure-identity>=1.15.0",
    "pydantic>=2.5.2",
    "pandas>=2.1.4",
    "numpy>=1.24.3",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.2",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "structlog>=23.2.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "flake8>=6.1.0",
    "pre-commit>=3.6.0",
]
ui = [
    "plotly>=5.17.0",
    "altair>=5.2.0",
]
search = [
    "duckduckgo-search>=3.9.6",
    "tavily-python>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-agent-project"
Repository = "https://github.com/yourusername/ai-agent-project"
Documentation = "https://github.com/yourusername/ai-agent-project#readme"
Issues = "https://github.com/yourusername/ai-agent-project/issues"

[project.scripts]
ai-agent-api = "app.api.main:main"
ai-agent-ui = "ui.streamlit_app:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "ui*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["app", "ui"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "streamlit.*",
    "langchain.*",
    "langgraph.*",
    "duckduckgo_search.*",
    "structlog.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app", "ui"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]